move test to more suitable location
